+++ Example/IntegrationTests/Sign/Helpers/ClientDelegate.swift
